Couldnt't change color of links with .navbar.scrolled
So added .navigation a infront of it to join both and didnt work either, instead only changed background and color of links.

Finally copied and created new style of navigation a and just removed the background and kept it same ,
and then removed .navigation a from the style which changed the navbar background color so that it applies to whole navbar.